Maybe... implemented .moveAI() to the UI

+++ src/main/kotlin/ui/Action.kt
@@ -2,5 +2,5 @@ package ui
 
 data class Action(
     val name: String,
-    val action: (Int) -> Unit,
+    val action: () -> Unit,
 )
